decompose markdown component
