+++ app/src/features/tickets/markdown/TicketMarkdownContent.test.tsx
@@ -1,8 +1,8 @@
-import { screen } from '@testing-library/react';
+import { screen, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { ITicketArtifact } from 'interfaces/useTicketsApi.interface';
 import { render } from 'test-helpers/test-utils';
-import { TicketMarkdownContent } from './TicketMarkdownContent';
+import { TicketMarkdownContent } from './TicketMarkdownContent/components/TicketMarkdownContent';
 
 const ticketArtifact = (
   artifact: Pick<ITicketArtifact, 'ticket_artifact_id' | 'artifact_id' | 'key'>
@@ -186,5 +186,61 @@ describe('TicketMarkdownContent', () => {
     );
 
     expect(screen.getByRole('button', { name: 'diagram.png' })).not.toHaveAttribute('href');
+  });
+
+  it('passes root Box props to the markdown container', () => {
+    render(<TicketMarkdownContent content="Ticket comment" data-testid="ticket-markdown-root" sx={{ mt: 1 }} />);
+
+    expect(screen.getByTestId('ticket-markdown-root')).toHaveTextContent('Ticket comment');
+  });
+
+  it('updates artifact link resolution when the artifacts prop changes', () => {
+    const { rerender } = render(
+      <TicketMarkdownContent content="[3ea79946-8cf2-4792-9239-5b148f9f95eb](/artifact/3ea79946-8cf2-4792-9239-5b148f9f95eb)" />
+    );
+
+    expect(screen.getByText('File not found')).toBeVisible();
+
+    rerender(
+      <TicketMarkdownContent
+        content="[3ea79946-8cf2-4792-9239-5b148f9f95eb](/artifact/3ea79946-8cf2-4792-9239-5b148f9f95eb)"
+        artifacts={[
+          ticketArtifact({
+            ticket_artifact_id: '3ea79946-8cf2-4792-9239-5b148f9f95eb',
+            artifact_id: '90b6df74-1b23-4064-ad62-f83c291d31d2',
+            key: 'tickets/ticket-id/upload/upload-id/report.pdf'
+          })
+        ]}
+      />
+    );
+
+    expect(screen.queryByText('File not found')).not.toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'report.pdf' })).not.toHaveAttribute('href');
+  });
+
+  it('scopes artifact lookup to the current component instance', () => {
+    const artifact = ticketArtifact({
+      ticket_artifact_id: '3ea79946-8cf2-4792-9239-5b148f9f95eb',
+      artifact_id: '90b6df74-1b23-4064-ad62-f83c291d31d2',
+      key: 'tickets/ticket-id/upload/upload-id/report.pdf'
+    });
+
+    render(
+      <>
+        <TicketMarkdownContent
+          content="[first](/artifact/3ea79946-8cf2-4792-9239-5b148f9f95eb)"
+          artifacts={[artifact]}
+          data-testid="first-markdown"
+        />
+        <TicketMarkdownContent
+          content="[second](/artifact/3ea79946-8cf2-4792-9239-5b148f9f95eb)"
+          data-testid="second-markdown"
+        />
+      </>
+    );
+
+    expect(within(screen.getByTestId('first-markdown')).getByRole('button', { name: 'report.pdf' })).toBeVisible();
+    expect(within(screen.getByTestId('second-markdown')).getByText('File not found')).toBeVisible();
+    expect(within(screen.getByTestId('second-markdown')).queryByRole('button', { name: 'report.pdf' })).not.toBeInTheDocument();
   });
 });
